Fixed HW4 DRM issues
Fixes issues on Apple TV+, DRM content on Safari, Apple Music etc.

Apple TV+ & Netflix confirmed working on my RX 580. I have shikigva as a child in the GPUs PciRoot but I don't think it makes a difference.

Thanks to Evan and Robert for bringing the fix to my attention!!

Co-Authored-By: blacksea1206 <63115499+blacksea1206@users.noreply.github.com>
Co-Authored-By: mlch911 <26565394+mlch911@users.noreply.github.com>